backend/requirements.txt
Click==7.0
Django==2.2.5
django-cors-headers==3.1.0
django-redis==4.10.0
dlib==19.17.0
face-recognition==1.2.3
face-recognition-models==0.3.0
mysqlclient==1.4.4
numpy==1.17.2
Pillow==5.2.0
pytz==2019.2
redis==3.3.8
sqlparse==0.3.0
uWSGI==2.0.18
